Refactor reduce scope of otherwise unused fixture
